Fix all qulice suggestions
