+++ pom.xml
@@ -77,11 +77,6 @@ SOFTWARE.
       <version>2.1.1</version>
     </dependency>
     <dependency>
-      <groupId>com.jcabi</groupId>
-      <artifactId>jcabi-http</artifactId>
-      <version>1.20.1</version>
-    </dependency>
-    <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-api</artifactId>
       <version>5.9.2</version>
