+++ src/main/java/com/github/lombrozo/testnames/rules/ml/ModelSourceFileSystem.java
@@ -1,22 +1,50 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2022-2023 Volodya
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 package com.github.lombrozo.testnames.rules.ml;
 
 import java.io.IOException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import opennlp.tools.postag.POSModel;
 
-public class ModelSourceFileSystem implements ModelSource {
+/**
+ * Model source from file system.
+ *
+ * @since 0.10
+ */
+public final class ModelSourceFileSystem implements ModelSource {
+
+    /**
+     * The path to the model.
+     */
     private final Path path;
 
-    public ModelSourceFileSystem() {
-        this("/tmp/opennlp/en-pos-perceptron.bin");
-    }
-
-    private ModelSourceFileSystem(final String file) {
-        this(Paths.get(file));
-    }
-
-    private ModelSourceFileSystem(final Path file) {
+    /**
+     * Constructor.
+     * Use path like "/tmp/opennlp/en-pos-perceptron.bin".
+     * @param file The path to the model.
+     */
+    ModelSourceFileSystem(final Path file) {
         this.path = file;
     }
